refactor: decouple the model from the eval pipeline
